1.12.2-1.0.1 Prebuild
initialized repository with contents of BonsaiTrees a MIT licensced open source project.

I noticed that Bonsai Trees will do crops if the configs are modified. I further noticed a user requested that be a default capability. The original mod author said he would not do it and that configs should be used.

I think that this would be a popular mod if this behavior was added as a default option. Since Bonsai Trees is actively maintained, rather that forking, i chose to create a new repository that will work for just crops.